Inverted speed meter mechanism, added randomized pace/breathe keys.
Instead of trying to run fast (which ultimately hurts you) you know have to attempt to maintain a steady pace. Gives much more of an incentive to try and press those keys. Randomized the keys needed to be input for both running and breathing! Next is added time stages that make the game incrementally harder (also balancing/ adding a timer to the breathe functionality).